The kadmin object must be recreated every time.
When the kadmin object is created, it generates a GSSAPI login packet to
authenticate to kadmind. This expires when our kadmin/admin ticket
expires, so we need to create a fresh kadmin object every time rather
than caching it.

Properly I should catch the appropriate GSSAPI exception and retry with
a new kadmin object. I don't believe this would ever be necessary, as a
kopf-level retry of a handler will get new credentials.

Since I need per-request context, move all the context to a per-request
object. This now makes it easy to pass logging through to the
kopf-provided logger, which logs onto the K8s object.