Pull the config from the ConfigDB.
For now we still need to patch the password into the fetched config
file. Once we can get an MQTT connection from the ServiceClient this
will no longer be necessary.

We now require Node UUID and Kerberos principal (username) in the
environment, and Sparkplug address in the config file. These are
redundant with each other and should all be looked up from our
credentials. (This is a problem across the board with all the clients.)