Replace the Tahu Sparkplug client.
The Tahu client insists on creating its own MQTT connection; this is no
good for us, as we want to get one from @amrc-factoryplus. Write a
replacement that accepts an already-connected MQTT client and runs a
Sparkplug client on top of it.

This is in JS not TS as this code wants to move into the library
eventually. All the Sparkplug nodes should use it. Probably most of
sparkplugNode.ts could be refactored into a generic MetricsNode class,
too.