Add a general Link schema.
This represents a link between this device and another object; another
Device, or potentially anything else identitified by a UUID. This is
modelled strongly on the RDF data model and on the <link> HTML element.

The 'meaning' of the link is specified by the Relation metric. These
Relations are UUIDs which should be registered in the ConfigDB. So, for
example, we could have a Relation UUID which means 'I provide power
monitoring information for the Target device'.