A proposed schema for Edge Agent metrics.
This schema is intended to be used for the Edge Agent Node metrics, and
gives information about the operation of the Edge Agent itself.

Config_Revision is intended to publish a UUID identifying a particular
revision of the configuration. This is not part of the config file
schema or generation yet but I think it should be, as it will allow an
Edge Deployment operator to monitor the current state of the Edge Agents
and poke those that aren't using their corrent current config.

The Alerts folder is to allow the Edge Agent to raise alerts. Once we
have the Edge Agent able to publish to Sparkplug given only its
credentials, it can raise alerts over the infrastructure for not being
able to read its config, or for problems with the southbound
connections. This should be easier to deal with than trying to monitor
container logs when data isn't flowing.

The UUIDs specified here for the Alert types are newly minted. They
would need registering in the ConfigDB before use.