Handle repos in groups.
Divorce repos from being so tightly linked to edge clusters: I don't
think this will be helpful. Give repos two-level names, and allow
permissions to be set on the group. This means we can create e.g. a
shared/ group that the edge clusters can pull shared config from, and we
can grant admins push rights to the whole group.

Repo groups are registered in the ConfigDB; this must be done
explicitly. Individual repos are also registered, but this should NOT be
done by hand; instead POST /git/:group/:repo should be called which will
create the ConfigDB entry and the git repo. Creating repos requires
Create permission on the group; it can't be granted on the repo yet.