Accept a different ccache for kadmin use.
A kadmin ticket must be obtained as initial credentials. If we are
becoming a service client we need a ccache with a TGT; this means we
need two k5start's running as sidecars, and so we need two ccaches. It
is simpler to pass the non-default ccache to kadmin than to the library.