Build clusters from a template.
Put the flux principal in auth groups to allow it to access shared
repos.

There is too much system-specific configuration finding its way in here,
especially when we start looking at granting permissions automatically.
Take a step back and push most of this into a template object in the
ConfigDB.

The create API has changed; the source links are now in the template,
and a template name is specified instead.

The template config format includes branch information which is not
currently used.