Accept JSON merge patch.
This is a rather limited patch format, but it has the useful property of
creating parent objects where necessary. JSON patch will not do this,
and also doesn't have a create-if-not-present operator, so patching in
an optional property is impossible without submitting multiple patches.

Check Content-Type so we can support different patch formats in the
future.