Update to the published utilities version.
Remove package-lock from Git and from the Docker build. It is causing
the Linux build to not include the GSSAPI libraries because package-lock
is generated on Windows where they are not available.

Adjust the Dockerfile to use a specific version of the base images.
Without this we will get incompatibilities when things change.