Update command args in influxdb-ingester.yaml
A line has been added to the command arguments in the influxdb-ingester.yaml file found within the templates/warehouse directory. This change was made to specify the use of the Node.js runtime and enforce the use of Node-style resolution for ES modules in the ingester.js file. This helps to eliminate possible discrepancies linked with the use of different module resolution strategies.