Add initScripts to values.yaml for influxdb
Init scripts have been added to the values.yaml file for influxdb setup. This includes the execution of a set of commands during the initial boot of the database to automate tasks like applying a template (influxdb2_operational_monitoring), adding jq (a lightweight and flexible command-line JSON processor), and creating a DBRP mapping. This changes are intended to streamline and automate the initial setup process, saving time and ensuring consistency.