Update instructions for data visualization
Updated the instructions in NOTES.txt template for users to understand the process of starting with Grafana after setting up devices and flowing data. Removed the instruction of logging into the InfluxDB instance as admin that was previously required. It's aimed to help users in building dashboards effectively in Grafana.