Update Helm tag version format
This commit modifies how the Helm package version is formatted in the GitHub actions workflow for publishing Helm charts. The change ensures that the 'v' prefix is stripped from the version tags. This conforms to the SemVer norm, thus maintaining consistency and compatibility with third-party tools expecting this standard format.