Add switchboard enable condition to Chart
This commit introduces a condition for the switchboard in the Helm chart. Previously, the switchboard component would always be deployed. However, with this change, we enable more flexibility by allowing the switchboard deployment to be toggled on or off via the 'switchboard.enabled' setting in the values.yaml file. The goal is to provide better control over the deployment based on the specific requirements of any given environment.