Create repos on the fly.
We need to change to being driven from the ConfigDB, so that repos can
be precreated as from a ConfigDB dump file. Autocreate a repo if the
ConfigDB entry exists. Store repos on disk under their UUID, so that
repos can be renamed without losing the link between object and storage.