Update logLevel in influxDB ingester to a configurable value
The value for the `LOG_LEVEL` in the `influxdb-ingester` configuration was hardcoded to `info`. This change replaces it with a value from `Values.warehouse.ingester.logLevel` making it configurable. With this change, we can easily modify the `LOG_LEVEL` without altering the codebase, just by changing a value in the environment configuration.