Update README and add flushInterval to warehouse ingester
A major update has been made to the README.md, replacing the extensive 'Values' section with a reference to the `values.yaml` file. This change is to ensure that the most accurate and up-to-date configuration parameters are always referenced directly from the `values.yaml`, which is the single source of truth.

Moreover, a new environment variable called `flushInterval` has been added into the Warehouse 'ingester' configuration on the `values.yaml` file and template. This change is designed to allow better control over data flushing, improving efficiency by allowing for fine-tuning of when buffered messages get sent, even if the buffer isn't full. We've also updated this component's version tag to reflect this new feature.