Disable cert-manager and external-dns installation in switchboard
This commit changes the configuration for 'cert-manager' and 'external-dns' under the 'switchboard' settings in values.yaml. The install settings for both these tools have been switched from 'true' to 'false'. This change implements a decision to not automatically install these components with the switchboard enabling, providing the user with more control over component installation.