Update repository paths for switchboard in Chart.lock and Chart.yaml
The repository paths for 'switchboard' in Chart.lock and Chart.yaml files have been updated. The change resulted in the removal of 'charts/' from all repository URLs and in updates to the 'digest' and 'generated' fields in Chart.lock. It seems the previous path was directing to an incorrect or non-existent location, and hence the paths were corrected to point to the right locations.