Return ETags.
I am going to document that ConfigDB entries always return a UUID as a
strong ETag. This makes a lot of things possible and simpler. Store
etags in the database, and return them on GET requests.

Don't return an ETag on a PUT response, as we have normalised the JSON.
(I am assuming that Postgres jsonb stringification produces consistent
results.)