Adjust the CRD.
Permit the class to be optional; in this case we will not create the
object or attempt to set the name. This is for Edge Agents where the
Manager has already set the Edge Agent UUID up in the ConfigDB; we are
just managing the Auth side of things.

Permit the account UUID to be optional; in this case the class is
required and we will create a new UUID. This is for accounts like the
Edge Sync where we are managing the whole account from scratch.