Separate out the account handlers.
Split the account handling into two more handlers: one to create the
object in the ConfigDB, and one to do the rest. This means that if we
succeed in creating a new object, that handler has entirely succeeded,
and we do actually record the new UUID in the annotation. Otherwise we
keep creating objects.