To be hashable we need to be immutable.
I should be able to make all my @fields objects immutable, but I have a
comment saying I wasn't able to make it work with inheritance. ACE does
not inherit and really does want to be hashable.

It seems we can't reasonable set fields in __init__ with immutable
objects. Use a factory method instead (this is arguably cleaner anyway).