Create a SparkplugNode CRD.
This is deployed by a Helm chart to inform us that we should be
monitoring a Node. The CRD supports three possibilities: a plain
Sparkplug Node which doesn't support F+, a Node registered by UUID in
the ConfigDB, and an Edge Agent which should be publishing the correct
schema and needs its config keeping up to date.