Split the node monitors into more classes.
Create a (currently mostly empty) base class for monitoring a general F+
Node; this could do monitoring useful for the Services, like checking
they are up and publishing the right sort of data. Create a subclass for
monitoring an Edge Agent which includes the existing config-reload
logic.

Expose a 'checks' sequence from the monitor object, which can be merged
into the which-nodes-are-we-monitoring sequence.