Push the list of hosts as an array.
This means merge-patch will update it as a single object rather than
merging in. For now, when we have a full list every time, this is
desirable.

Possibly these hosts should be exposed in the ConfigDB as objects of
their own; then we could have hosts moving between clusters. (This
machine is plugged into this cell gateway; I don't care about clusters,
just make the EA run there.)

Remove the PATCH-then-PUT logic. The ConfigDB handles this now.