Add a shareLatest operator.
This is like UniRx's ReadonlyReactiveProperty: it makes a sequence where
you can always subscribe to get the latest value, followed by any
updates. The correct options to achieve this are not obvious, so make an
operator.

One of the problems I am trying to solve with this is the problem of
sequences not being disposed properly when noone is interested in the
results. In particular, if we are trying to get a BIRTH from a device,
and making rebirth requests, we don't want to carry on doing that if our
subscriber has gone away. However, sometimes we get no subscribers as a
transient condition, when our only subscriber is retrying and will
resubscribe very soon.

If our source sequence has stopped we must reset our Subject: we have
potentially missed updates. If it's expensive to start again then that's
just tough. So allow the source to remain subscribed with no consumers
for a (fairly arbitrary) 1 second, and then reset. I don't think we can
do better than that; Rx doesn't have 'I am unsubscribing but I'm still
interested really and will be coming back soon' as a primitive.