Make sure the birth sequence can be cancelled.
We were creating a permanent subscription to a sequence attempting to get
a birth certificate by issuing rebirth requests. This meant that the Edge
Monitor was getting stuck when a failed EA deployment was removed: the
EA was never going to come online, and the EM had unsubscribed, but we
were still issuing rebirth requests.