Cache a shared sequence for each topic
Create a sequence watching a specific topic, with subscribe/unsubscribe
logic attached to sub/unsub to/from MQTT. Share and cache this sequence
so we accumulate consumers in the share refcount, and our MQTT
subscriptions only change when they need to.

The sequences exposed to users use rx.defer to ensure we always use the
latest cached topic sequence. This allows us to safely destroy unused
topic sequences without risk of them being resubscribed to.