We were not emitting metrics from the first BIRTH
The process of resolving aliases was naive: it was attempting to resolve
aliases on BIRTH as well as DATA, and so didn't start emitting packets
until it had seen two BIRTHS.

Births do not use aliases. Resolving aliases wrt the previous birth is
simply incorrect.

This has changed the format of the payload of SparkplugDevice.births,
but noone cares. A Map of metrics in the birth would still be useful,
but keeping the structure the same as DATA makes this processing easier.