Add an Action to do a full build. (#3)
It doesn't seem to be possible to usefully push Java builds anywhere.
The GH Maven repo can only be accessed with a PAT, even from actions on
other repos. Maven Central would be difficult to set up access to.

Bring in the Java code as a submodule. Adapt the build to run Maven and
make the build zipfile available to Docker.