Support getting client creds with a keytab
For reasons I don't entirely understand, using k5start to fetch our
client creds is not working properly. It appears that the Java GSS
library is getting an initial TGT from the provided ccache but ignores
updates after that.

Possibly this could be fixed with different config options, but for now
try just using the keytab directly instead.