Create new GSS credentials each time
The GSSAPI documentation implied to me that a GSSCredential was just a
handle pointing to a ccache or a keytab, and so it could be treated as
constant. I am having issues with the client not renewing its TGT, so
perhaps that isn't the case. Try creating a new credential every time we
need one.