Login to Kerberos again when our ticket expires
Java is doing its own ccache management, using private ccaches stashed
inside Subjects. This means we need to check the lifetime on the
credentials and renew them / log in again as needed.

This requires a fairly major rework of the code. I had been assuming
that once we'd got a GSSCredential that held a handle to a keytab or a
ccache and would pull credentials as needed; I think this is what the C
library does. Certainly from Javascript I just call createClientContext
and let the library handle finding the ccache and making sure we have
the tickets we need.