Create the Helm chart repo and the repo groups
Define an 'EDO Config' application holding the UUIDs the EDO needs to do
its job. (The EDO doesn't read this yet, but it needs to.)

Create repo groups for the edge clusters and the shared repos; these
repo groups are now just Auth groups, with no significance to the Git
service. Create a repo for the edge helm charts.

Ensure that if ConfigDB entries for any of these already exist we don't
replace them.