Mirror the helm-charts repo from Github
Start with a ref on the Github edge-helm-charts repo. This is configured
in the ACS Helm values file so we can control what we pull. Pull that
ref to branch `acs` in our local helm charts repo. Then, iff `main` was
equal to `acs` before we started, also update `main`.

This is all still EDO setup. I would like to move these setup pieces
into the repos they belong to, and pull them back here as submodules.