Update service dumps
This does not attempt to handle the dynamic groups created by
service-setup; there is very little point in any case since a limited
ManageGroups doesn't work correctly.

This does not grant permission to watch the ConfigDB because we can't do
that with fixed UUIDs yet. Grant explicitly in the ACS dumps for now.