Make sure we create the default branch
If we clone an empty repo the server doesn't tell us what it thinks the
default branch is called. The git client then defaults to 'master',
which is wrong. The next time we clone we get an error because HEAD
doesn't exist (why we don't get that error for an empty repo I don't
really understand).

Make sure we switch to the right branch. For now assume this is 'main';
I don't really like this but there is no standard way to push a default
branch to the server.