Create an endpoint to return a bootstrap URL
Currently the bootstrap URLs are fixed and public. This makes me a
little nervous; I would like to include an expiring token in a URL which
doesn't need auth. So provide an endpoint which can be used by the
Manager to retrieve a bootstrap URL for a cluster. Currently this just
returns the fixed public URL, but I hope that will change.

This is POST because it might generate a new token server-side. The
returned URL is relative to the request URL, to avoid trying to find a
suitable external address.