Load dumps into the services
Going forward it is going to be better to load JSON dumps once at ACS
deploy time rather than every time the service starts. This will make
service startup much quicker and also gives the administrator more
control over when dumps are reloaded.

Create a ss action to load in a set of JSON dumps built into the image.
Include dumps for the ConfigDB entries we need to do our job, and some
of the new entries required by the Clusters service.