Find our address from our credentials
This means the Directory MQTT client will not start up until after the
Auth and ConfigDB have started up and been initialised, but that should
be OK. Both those services are registered over HTTP now.

We can't quite (yet) avoid needing to know our UUID in the Helm
template. The HTTP ping response includes the MQTT device ID; this has
always been there for linking purposes. I am inclined to just remove
this link; nobody uses it, and except for the Directory itself we always
get hold of an HTTP URL and a Sparkplug Device UUID at the same time,
from the Directory.