Check for service override permission first
When the Auth service is starting up we run an initContainer, as
ROOT_PRINCIPAL, to register the Auth service URL. If this registration
request makes any Auth service requests they will time out as the Auth
service isn't running yet.

Kerberos checks for ROOT_PRINCIPAL don't go to the network, so we need
to make sure these are performed first if they're going to be.