Service registrations are owned by devices
I was making the system over-complicated with registration owners
separate from registered MQTT devices. The normal case is a principal,
which may or may not be a Sparkplug Node, registering a service URL. In
this case the second Directory request to find the principal's Sparkplug
address will only succeed if the principal has published a BIRTH.

The case of a Sparkplug Device advertising a service is still minimally
supported, but any HTTP registration would need to use the setid
endpoint. (This could be granted automatically to Nodes for their
Devices.)

There is a security hole here: looking up a service device UUID via the
Directory is vulnerable to spoofing. This has always existed. It could
be avoided in cases where we know the service is provided by a Node (the
normal case) by looking in the ConfigDB instead. Or the Directory could
check permissions before indexing.