Stop using a Helm Template for the bootstrap
We were using the edge cluster helm template to find the domain and
realm information the bootstrap script needs, on the theory that this
would keep consistency with the deployed helm chart. But with a choice
of cluster helm chart available, and the additional risk of needing to
keep the edge cluster values in sync with our requirements here, it's
just not worth it. Pull values from our own environment instead.