Throw errors when we can't contact services
Currently we run into a problem when the upstream services are down:
clients are permitted to connect but not granted any ACLs. The client
generally doesn't realise there is a problem and so doesn't attempt to
reconnect.

This will cause authentication failures in this case. A Server
Unavailable error would be better.