Retry on device lookup failure
If we look up a Device and it isn't in the Directory, or a Node and it
isn't in the ConfigDB, retry with a delay instead of returning an
invalid address.

We were entering a fast loop and consuming memory; I think this.packets
was failing and so the retry logic on the rebirths was skipping the
timeout.