See if this fixes the build
For some reason the build of v1.0.2 is failing. GH Actions is not giving
me much information about why; the Docker metadata step is failing to
pick up the tag, but it's not clear why.

The only difference between this workflow and some of the others that
work is this. See if it makes a difference.