Unlink the file but hold open while we write
We are falling into an annoying bug in isomorphic-git where if a certain
subset of lstat information hasn't changed, in particular inode and
timestamps-to-the-second, then it concludes the file can't have changed
either. This is incorrect when we clone-edit-commit without changing the
size of a file.

Unlink the old file but hold it open over the write. This will force a
new inode to be used, making a change to the lstat result.