Don't publish when we're offline
I thought the MQTT library handled it and queued the packets for later
delivery, but apparently not, or maybe this just doesn't apply when we
haven't connected at all yet. We only care about publishing the current
state, so just skip updates when we're offline; we'll birth when we go
online in any case.