Remove the old metric builder functions
This includes a lot of metrics to support command escalation over MQTT.

These are still used by the services to generate their Sparkplug packets
but they were always a bit of a hack. Either they can stay in utilities
as compatibility until they are replaced or they can go out to a
separate package.