Support connecting to multiple brokers
This is only supported with cross-realm auth. Accept multiple Directory
URLs and connect to the broker from each. Merge all the packets into a
single graph, but make sure we look up devices in the correct Directory.

Icons all come from the first ConfigDB, with the caching it's too messy
to try to avoid that. Decouple the MQTT client from the Icons object.