Fix up a refactoring error from account creation
At some point this code was doing both key and account creation, and
returning a combined status object. But the refactoring was incomplete
and has led to HAS_OLD_KEYS being set in error.

As the account creation is now handled entirely separately just put this
back to what was there originally.