Forward compat for an API change
Starting to split up the `utilities` package has raised some issues with
the API. This method should always have returned a Promise; without this
it is impossible to dynamically load the Sparkplug support.

The `js-service-client` package has changed the API to return a Promise.
The `utilities` package will soon be importing its ServiceClient from
there. Fortunately JS allows you to await a non-Promise.