Work around a problem with Rollup's JS parser
When the Dockerfile runs `bun run build`, Rollup is transpiling the JS.
There appears to be a bug in Rollup's JS parser such that it inserts a
semicolon after `await` at the end of a line; this is incorrect and
breaks the ServiceClient code. For now hack the module after
installation to get this building.