Don't delete Alerts or Links
Add a 'stale' field instead; this is equivalent to the GI `deleted`
property in the ConfigDB. I really need a unified solution here.

Objects which are 'stale' do not appear in any API results. For now I am
not attempting to track history, but if we delete Alert records then the
change-notify code fails.