Display Links for Alerts
These include useful information such as the Node monitored for Monitor
Alerts.

It's not obvious how to stringify the UUIDs used in Links. Obviously
where there is a GI name that is correct; but for many links there won't
be. For now, try the GI name, and then look for a Sparkplug Address (via
the Directory, to catch Devices).