Switch from Nginx to Node.js serve activity monitor
In the Dockerfile, the base image has been changed from Nginx to Node.js. The configuration files specific to Nginx have been removed and replaced with 'serve', a simple Node.js static file server. This removes the overhead of having to configure Nginx for a simple static file serving scenario.