Rework rebirth logic
Waiting for packets with rx.timeout() has the problem that once you've
timed out you stop watching for packets. So if a packet comes in during
our jitter delay, we won't see it. It also causes a lot of subscription
and unsubscription if nothing else is watching the device.

Use switchMap instead, like the offline logic.