Stop sending repeated rebirths
In general it is not the responsibility of our consuming application to
try to wake a Device up. That is the infrastructure's responsibility. So
we should not be sending multiple rebirth requests.

We send a single request when we first need a BIRTH. If the device
responds, we have what we need; if not, it will BIRTH when it wakes up.
We do still need to retry in the case of cmdesc failure, as we may not
be able to contact the cmdesc service at all, or something.

This is technically an API change; SparkplugDevice.rebirth() now returns
an Observable rather than a Promise. But we are still on 0.0.x, this was
an internal method until I needed it in the Monitor, and the Monitor is
the only consumer and doesn't care about the change.