Add fullscreen mode in acs-admin application
The View and layoutStore in the acs-admin application were modified to support fullscreen mode. A toggle functionality was added in the layout store, and extra conditional statements were included in the view to adjust the UI when the fullscreen mode is activated. A keyboard shortcut (escape key) was also implemented to quickly toggle between fullscreen and normal mode.