Run a broker for drivers to connect to
We are using the Aedes broker library. Create a broker and perform auth
on clients.

The broker URL to listen on and the location of the passwords for the
drivers are taken from the environment rather than from the config as
these will need to be synchronised with the driver deployment. This
means they are Helm values rather than edge agent config items.