Add address groups and driver polling
OPCUA and MTConnect support automatic polling within the protocol. S7
supports group queries. It is a good idea to allow drivers to be
efficient where they can be.

Also support an overall `active` topic for a driver, indicating whether
the relevant Edge Agent connection is online. This also serves as a
prompt for the driver to republish its status when the Edge Agent starts
up.