Deploy drivers in the edge agent pod
Given that all drivers are deployed identically, and that in general we
want to run them on the same host as the edge agent, deploy them as part
of the edge agent pod and from the same Helm chart.

This does not preclude external drivers, but would need additional
configuration to set up external IPs and to listen to the wildcard
address.