Make Handler.connect return a Promise
The callback API for the handler to report connection status to the
driver is not useful. Connection problems for TCP-based protocols are
only likely to occur on the initial connection attempt, so make
`connect` return a Promise to a status string. If the connection is
closed unexpectedly then we delay and attempt to reconnect.

Handling RESTish protocols like this might be tricky, but it's not
obvious how to map REST connection problems to Sparkplug at all.