Stop naming K8s resources with UUIDs
It is getting very irritating trying to work out which pod is which when
they are all named with UUIDs. We are also running into k8s name length
limits when k8s extends the Deployment name to create ReplicaSet and Pod
names.

Our `name` must be unique in any case, or everything breaks. Use a
sanitised form of the name in our k8s resource names.