Change Makefiles to use docker buildx
Use of buildx appears to be mandatory now. Even an explicit `docker
builder build` seems to invoke buildx. Adjust the Makefile to suit; this
means the `build` target now performs `push`, and we need `pull` target
for cases where we want an up-to-date local image.

Support multi-platform builds, now we can.