Refactor service-setup template
Refactor the service-setup template to use a generic list of steps. This
allows users to customise the service-setup actions.

The individual steps are defined as a dict in the values file, and then
the choice of which steps to run and in what order is made separately.
This is because the json-merge algorithm overwrites array values
wholesale, so with the step definitions inside the array any
customisation would lose the stock step definitions.