List schemas out of the ConfigDB
There is no need to keep a copy of the schema list in the Manager
database at all, we can just go direct to the ConfigDB. Pull the
metadata we need for the 'choose a schema' overlay.

The frontend logic has changed: instead of making individual calls we
now retrieve all the data and filter client-side. Given the relatively
small number of schemas there are likely to ever be this si unlikely to
cause a performance problem.

The search function is now broken because that relied on the backend
doing the search via Melisearch. This could be reimplemented on the
frontend but would require messing with the ColumnList component.