Update edge cluster bootstrap
Operator correctly when our namespace already exists. Fail correctly
when the user types the wrong password; this is an easy error and should
not cause damage.

Don't attempt to install Flux if it is already there. This is a
different decision from the 'bare deployment' option in the Manager,
which controls deployment of Flux manifests for Flux itself and for
Sealed Secrets. A bare deployment where Flux is not already installed
will install Flux once but not update it; a non-bare deployment to a
cluster that already has Flux will attempt to use the existing Flux for
the initial bootstrap and then upgrade to a self-updating setup.