Add a publish interval per-device
Currently the Edge Agent accepts a pubInterval at the Node level to
control buffering of metrics. This is difficult to create from the
Manager as we don't have a UI to edit properties of a Node. It is also
inflexible; this should be a Device configuration anyway.

Make this available as a per-Device config option. This will require
Edge Agent changes.