Spec a new notify protocol
I don't think notify over Sparkplug works at all. There are just too
many moving parts: we have to adjust MQTT ACLs, the client has to deal
with connecting to MQTT, we have to either refer back to the HTTP API or
work out how to represent HTTP resources as Sparkplug data...

Rework entirely as a generic JSON API change-notify interface over a
WebSocket. The client will end up with a WS open to each service it is
watching; I don't think this is a problem.