Define and process an auth exchange on the WS
The Web API WebSocket interface does not have any options to set HTTP
headers. While a browser will probably supply default credentials (auth,
cookies) this will not help with a ServiceClient that needs to perform
its own credential management; nor will it help from Node.

Specify an initial auth exchange to deal with this. The exchange is
deliberately minimal and requires the client to get a token from the
service via the normal HTTP API.