Implement notify subscriptions
Currently this only allows subscrition to an individual config entry;
the protocol will need extending to allow multi-entry subscriptions
before this will be very usable. Change the UPDATE method to WATCH in
the spec to better reflect its purpose; I think I will need to introduce
SEARCH in future for a subscription to child URL updates.