Handle nonexistent apps
I'm not sure what will happen if an app is deleted... I'm not sure
that's a case worth handling, though.

This means reworking search_full as a plain Promise-returning function,
and adding an `asyncState` operator to accept Promises to state. I don't
think trying to force these Promises into Observables would be helpful;
something like RxJava's Single would be best, but with only Observable
there is always the question of how to handle multiple results.