Allow clients to receive HTTP statuses
While in many circumstances we are only interested in the bodies of
responses, and it's reasonable to suppress errors, this is not always
the case. Define a Maybe monad to handle the three important cases: OK,
not-found, and other-errors. Define `_full` methods to return complete
responses wrapped in a Maybe.