Simplify the object model a little
I am not trying to implement a generic triplestore here, I am
implementing the F+ ConfigDB. Restrict our object model to fixed ranks
of objects and enforce this in the database. Stop trying to deduce rank
information from membership and subclass and store it directly. Remove
the Object/Class/Rank top-level objects and just leave the ranks at the
top of the structure for now.