Allow external edge driver deployment
Without this option it isn't possible to deploy an external driver as
the credentials are not created. Long-term we need a better option for
getting hold of the credentials than reading them from the cluster;
probably setting them via the sealed-secrets path. However this would
mean having independent Secrets for on-cluster and external drivers.