Keep special app configs in the database
Refactor the special app handling so that their configs are stored in
the database, and this is their canonical location. The additional
storage is treated as a 'sideways' cache.

The purpose of this is to implement PATCH on Registration, but for now
leave that alone and just port the config schema over. This removes a
lot of special case code for config schemas; all we now do is cache the
schemas as they go by.