Accept updates to primary class and rank
Changes of primary class will add the object as a member of the class if
necessary. Existing class membership is unaffected. This will be needed
when loading dumps.

For now only allow updates of rank to or from Individual, which also
implies the updated object must have no subclasses or members. The full
problem is more difficult and I am deferring it for now.