Emit notify updates for class changes
The internal sequence just sends an empty update for class changes, and
the subscription handler has to deal with looking up the new contents of
the class in question. There is a small risk here we will miss an update
but any state published will be the most recent.

Ensure the lookup endpoints return distinct lists.