Set service-setup backoff limit very high
There is no reason to ever stop trying. We cannot say 'don't stop', this
can only be achieved via `podFailurePolicy` which requires
`restartPolicy: Never` on the Pod, and will create a new Pod for every
new attempt. This will re-run all the containers, including those that
succeeded.

We may need to revisit this list-of-initContainers idea. Perhaps
multiple Jobs would be better? But then they will need to wait for each
other as appropriate. With non-atomic services just trying and failing
may leave a mess.