Database functions to load object from a dump
The consistency checking depends critically on the existing objects in
the database so it's going to be easiest to just get the data into the
database and process it there.

This only handle the object definitions. The configs can still be loaded
separately.

Create an SQL function to validate our invariants where they can't be
checked with FKs. We should probably call this at startup.