Implement the new auth class structure
Some revision of the structure was needed.

This omits the definitions of the composite permissions, as these are
enumerated classes (they have a fixed known list of members) which we
don't support yet in v2 dumps.

I have included REQUIRE comments to indicate deps between the dumps.
With the current arrangements it isn't possible to make a dump entirely
self-contained; the design is that we rely on existing objects to
determine the ranks of the new objects. These comments are currently
ignored by service-setup.