Handle existing objects when we re-rank
When the rank of an object changes we need to handle existing objects.
Currently the design is:

* All ranks are reset based on primary class.
* Parent relations from existing objects which conflict with the new
  ranks will raise an error.
* Existing objects promoted from individual to class are subclassed
  directly under the appropriate rank superclass.

I'm not sure the second of these is the best behaviour, but I think it's
safest for now. Silently removing the invalid relations would be the
other possibility, which is less likely to cause an ACS upgrade to fail
but more likely to lose something important.